Move empty dispatch handlers into a dedicated module
